+++ src/pages/about-page/layout/about-general/about-general.tsx
@@ -1,17 +1,14 @@
 import { type FC } from 'react'
 import { Helmet } from 'react-helmet-async'
 
-import { useBreakPoint } from 'src/hooks/useBreakPoint/useBreakPoint'
-
 import { BlockquoteSection } from 'src/pages/about-page/layout/about-general/components/blockquote-section/blockquote-section'
 import { GallerySection } from 'src/pages/about-page/layout/about-general/components/gallery-section/gallery-section'
 import { DescSection } from 'src/pages/about-page/layout/about-general/components/desc-section/desc-section'
-import { MobileTextContainer } from 'src/components/mobile-text-container/mobileTextContainer'
+import { CollapsibleText } from 'src/components/collapsible-text/collapsible-text'
 
 import styles from './index.module.scss'
 
 export const AboutGeneral: FC = () => {
-	const breakpoint = useBreakPoint()
 	return (
 		<div className={styles.aboutGeneralPage}>
 			<Helmet>
@@ -21,11 +18,7 @@ export const AboutGeneral: FC = () => {
 			<h2>Атманов угол</h2>
 			<BlockquoteSection />
 			<GallerySection />
-			{breakpoint !== 'M' ? (
-				<MobileTextContainer item={<DescSection />} lineClamp={22} />
-			) : (
-				<DescSection />
-			)}
+			<CollapsibleText item={<DescSection />} lineClamp={22} collapsePoint={'M'} />
 		</div>
 	)
 }
